test(variant): drop the reflective inferrer probe and share the inference helpers
The trait probed for VariantShreddingRuntime reflectively so it compiled
before #18961 landed; it calls lookupInferrer directly now. The three
inference assertions TestVariantDataType carried (assertInferredTypedValue,
assertInferredTypedValueIn, typedValueNonNullCount) move into the trait
so both suites share them.
