+++ hudi-spark-datasource/hudi-spark/src/test/scala/org/apache/spark/sql/hudi/dml/schema/TestVariantDataType.scala
@@ -36,8 +36,6 @@ import org.apache.hudi.testutils.DataSourceTestUtils
 import org.apache.hudi.testutils.HoodieClientTestUtils.createMetaClient
 
 import org.apache.hadoop.fs.{Path => HadoopPath}
-import org.apache.parquet.hadoop.ParquetFileReader
-import org.apache.parquet.hadoop.util.HadoopInputFile
 import org.apache.parquet.schema.Type
 import org.apache.spark.sql.{Row, SaveMode}
 import org.apache.spark.sql.catalyst.TableIdentifier
@@ -1439,43 +1437,6 @@ class TestVariantDataType extends HoodieSparkSqlTestBase with VariantShreddingTe
   }
 
   /**
-   * Pins an INFERRED shredding layout of `column` across every data parquet file of the table:
-   * see [[assertInferredTypedValueIn]].
-   */
-  private def assertInferredTypedValue(tablePath: String, column: String, leg: String,
-                                       present: Seq[String], absent: Seq[String] = Seq.empty): Unit = {
-    assertInferredTypedValueIn(listDataParquetFiles(tablePath), column, leg, present, absent)
-  }
-
-  /**
-   * Pins an INFERRED shredding layout of `column` in the given parquet files: the variant group is
-   * shredded, carries the VARIANT logical type (the inferrer only exists on Spark 4.1+, whose
-   * parquet ships the annotation), its typed_value has every `present` member and none of
-   * the `absent` ones (e.g. an avro-illegal key the inferrer dropped), and typed_value holds
-   * non-null values per the block column statistics.
-   */
-  private def assertInferredTypedValueIn(files: Seq[String], column: String, leg: String,
-                                         present: Seq[String], absent: Seq[String] = Seq.empty): Unit = {
-    assert(files.nonEmpty, s"[$leg] should have at least one data parquet file")
-    files.foreach { filePath =>
-      val variantGroup = getFieldAsGroup(readParquetSchema(filePath), column)
-      assert(variantGroup.containsField("typed_value"),
-        s"[$leg] $column should be shredded with an inferred typed_value. File: $filePath Schema:\n$variantGroup")
-      assert(Option(variantGroup.getLogicalTypeAnnotation).exists(_.toString.contains("VARIANT")),
-        s"[$leg] inferred variant group must carry the VARIANT logical type. Schema:\n$variantGroup")
-      val typedValue = getFieldAsGroup(variantGroup, "typed_value")
-      present.foreach(member => assert(typedValue.containsField(member),
-        s"[$leg] inferred typed_value should contain $member. Schema:\n$typedValue"))
-      absent.foreach(member => assert(!typedValue.containsField(member),
-        s"[$leg] inferred typed_value must not contain $member. Schema:\n$typedValue"))
-      // The footer schema alone cannot tell real shredding from a file where every row fell
-      // back to the residual value with typed_value all-null.
-      assert(typedValueNonNullCount(filePath, column) > 0,
-        s"[$leg] inferred typed_value of $column should hold non-null values. File: $filePath")
-    }
-  }
-
-  /**
    * Block types of every log block in the table, read from the log files themselves. Tests that pin
    * a log format assert on this rather than on file names: native logs carry a .log.parquet suffix,
    * but an inline log file is named the same whether its data blocks are avro or parquet.
@@ -1497,34 +1458,6 @@ class TestVariantDataType extends HoodieSparkSqlTestBase with VariantShreddingTe
       } finally {
         reader.close()
       }
-    }
-  }
-
-  /**
-   * Sums the non-null value counts of the typed leaf columns under `column`.typed_value across
-   * all blocks of the file, from the block column statistics. The residual `value` leaves of
-   * shredded object fields are left out: they hold the rows that fell back, so counting them
-   * would let a file with every field fallen back pass as shredded.
-   */
-  private def typedValueNonNullCount(filePath: String, column: String): Long = {
-    val conf = spark.sparkContext.hadoopConfiguration
-    val inputFile = HadoopInputFile.fromPath(new HadoopPath(filePath), conf)
-    val reader = ParquetFileReader.open(inputFile)
-    try {
-      val prefix = s"$column.typed_value"
-      reader.getFooter.getBlocks.asScala.flatMap(_.getColumns.asScala)
-        .filter { c =>
-          val dot = c.getPath.toDotString
-          (dot == prefix || dot.startsWith(prefix + ".")) && !dot.endsWith(".value")
-        }
-        .map { c =>
-          val stats = c.getStatistics
-          // A chunk without a written null count must not pass as holding non-null values
-          if (stats == null || stats.getNumNulls < 0) 0L else c.getValueCount - stats.getNumNulls
-        }
-        .sum
-    } finally {
-      reader.close()
     }
   }
